// 07-MoviesApp/src/infrastucture/interfaces/movie-cast.responses.ts
// Generated by https://quicktype.io
export interface MovieDBCastResponse {
    id: number;
    cast: MovieDBCast[];
    crew: MovieDBCast[];
}

export interface MovieDBCast {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: null | string;
    cast_id?: number;
    character?: string;
    credit_id: string;
    order?: number;
    department?: string;
    job?: string;
}